backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
spacy==3.7.2
sentence-transformers==2.2.2
huggingface_hub==0.23.4
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
